Fix race condition in Search.
Change the Execute() method of Search so that it does not modify internal
state of Search. Before Execute would change the SearchParams.Limit member
of Search to 100 it it was 0. Now Execute makes a copy of SearchParams and
changes Limit in the copy which is thread-safe.

Also changed NewSearch() to make a defensive copy of SearchParams.